+++ src/main/java/com/tictac/demo/service/CompetenciaService.java
@@ -17,12 +17,34 @@ public class CompetenciaService {
     public Optional<Competencia> getCompetencia(Integer id){
         return competenciaRepository.findById(id);
     }
+
     public Competencia saveCompetencia(Competencia competencia){
-        return competenciaRepository.save(competencia);
+        if(competencia.getNombre() == null || competencia.getNombre().trim().isEmpty()){
+            return null;
+        }else{
+            return competenciaRepository.save(competencia);
+        }
     }
 
-    public void deleteCompetencia(Integer id){
-        competenciaRepository.deleteById(id);
+    public String updateCompetencia(Competencia competencia){
+        if(competenciaRepository.existsById(competencia.getIdCompetencia())){
+            Optional<Competencia> compe = competenciaRepository.findById(competencia.getIdCompetencia());
+
+            compe.get().setNombre(competencia.getNombre());
+            competenciaRepository.save(compe.get());
+            return "Competencia actualizada con éxito";
+        }else{
+            return null;
+        }
+    }
+
+    public String deleteCompetencia(Integer id){
+        if(competenciaRepository.existsById(id)){
+            competenciaRepository.deleteById(id);
+            return "Competencia eliminada con éxito";
+        }else{
+            return null;
+        }
     }
 
     public List<Competencia> listCompetencia(){
